+++ sales-tax/src/main/java/com/complyt/business/nexus/data_extractor/AmountExtractor.java
@@ -1,5 +1,7 @@
 package com.complyt.business.nexus.data_extractor;
 
+import java.math.BigDecimal;
+
 public interface AmountExtractor {
-    float extract();
+    BigDecimal extract();
 }
