+++ sales-tax/src/main/java/com/complyt/domain/sales_tax/RatesMetaData.java
@@ -1,4 +1,6 @@
 package com.complyt.domain.sales_tax;
 
-public record RatesMetaData(double cityDistrictRate, double countyDistrictRate) {
+import java.math.BigDecimal;
+
+public record RatesMetaData(BigDecimal cityDistrictRate, BigDecimal countyDistrictRate) {
 }
