+++ sales-tax/src/main/java/com/complyt/domain/sales_tax/SalesTax.java
@@ -2,6 +2,8 @@ package com.complyt.domain.sales_tax;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @With
-public record SalesTax(double amount, SalesTaxRates salesTaxRates) {
+public record SalesTax(BigDecimal amount, SalesTaxRates salesTaxRates) {
 }
